2 training debug (#1)
* [~] wip

* [~] fixes

* [+] added comet experiment

+++ src/pl_modules.py
@@ -96,7 +96,8 @@ class BasicClassificationModule(pl.LightningModule):
     def training_step(self, batch: Tuple[Tensor, Tensor], batch_idx: int) -> Tensor:
         x, y = batch
         y_hat = self(x)
-        loss = self.loss(y_hat, y)
+
+        loss = self.loss(y_hat, y.float())
 
         self.log(f"train_loss_{get_loss_name(self.loss)}", loss)
         if self.sched is not None:
@@ -108,7 +109,8 @@ class BasicClassificationModule(pl.LightningModule):
     ):
         x, y = batch
         y_hat = self(x)
-        loss = self.loss(y_hat, y)
+
+        loss = self.loss(y_hat, y.float())
         self.log(f"val_loss_{get_loss_name(self.loss)}", loss, sync_dist=True)
 
         y_hat = torch.sigmoid(y_hat)
